requirements.txt
absl-py==1.3.0
contourpy==1.0.6
cycler==0.11.0
fonttools==4.38.0
kiwisolver==1.4.4
matplotlib==3.6.2
numpy==1.24.0
packaging==22.0
pandas==1.5.2
Pillow==9.3.0
pkg_resources==0.0.0
pylibxc2==6.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
scipy==1.9.3
seaborn==0.12.1
six==1.16.0
yapf==0.32.0
